requirements.txt
# Discord
discord.py[voice]>=2.3.2
PyNaCl>=1.5.0

# Transcription
faster-whisper>=1.0.0
torch>=2.0.0
torchaudio>=2.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0

# Vector Store (for future topic analysis)
qdrant-client>=1.7.0

# Embeddings (for future)
sentence-transformers>=2.3.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Audio processing
numpy>=1.24.0
